Updated web site for 4.1 launch.
git-svn-id: https://svn.code.sf.net/p/freemat/code/trunk@4241 58e251e6-400e-0410-b350-93cf1ec51ee9

+++ FreeMat/tests/jit/tjit6.m
@@ -1,5 +1,5 @@
 function tjit6
-A = int32(ones(1,10000));
+A = ones(1,10000);
 k = 0;
 for i=1:10000
   k = k + A(i);
